refactor(subscriptions): split lifecycle extension logic
