+++ backend/bot/services/subscription_service_impl/lifecycle.py
@@ -5,12 +5,14 @@ from db.dal import payment_dal, subscription_dal, tariff_dal, user_dal
 from ._typing import SubscriptionServiceMixinContract
 from .lifecycle_activation import SubscriptionLifecycleActivationMixin
 from .lifecycle_details import SubscriptionLifecycleDetailsMixin
+from .lifecycle_extension import SubscriptionLifecycleExtensionMixin
 from .lifecycle_panel import SubscriptionLifecyclePanelMixin
 from .lifecycle_switch import SubscriptionLifecycleSwitchMixin
 
 
 class SubscriptionLifecycleMixin(
     SubscriptionLifecycleActivationMixin,
+    SubscriptionLifecycleExtensionMixin,
     SubscriptionLifecycleDetailsMixin,
     SubscriptionLifecycleSwitchMixin,
     SubscriptionLifecyclePanelMixin,
